# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gateway",
        "description": "Swagger do Gateway",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/auth/signup": {
            "post": {
                "tags": [
                    "gateway"
                ],
                "summary": "Criar um novo usuário",
                "description": "Criar um novo usuário",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).+$"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nome": "a3fb6",
                                    "email": "teste@teste.com.br",
                                    "password": "123456"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário criado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "gateway"
                ],
                "summary": "Login do usuário",
                "description": "Login do usuário",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).+$"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "teste@teste.com.br",
                                    "password": "123456"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "nome": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "created_at": {
                                                    "type": "string"
                                                },
                                                "updated_at": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "properties": {
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                },
                                                "expires_in": {
                                                    "type": "string",
                                                    "example": "31535999"
                                                },
                                                "access_token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLC"
                                                },
                                                "refresh_token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLC"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "E-mail não existe"
                    },
                    "401": {
                        "description": "Senha inválida"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "gateway"
                ],
                "summary": "Cancela o token do usuário",
                "description": "Cancela o token do usuário",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário deslogado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/auth/user": {
            "get": {
                "tags": [
                    "gateway"
                ],
                "summary": "Retorna os dados do usuário",
                "description": "Retorna os dados do usuário",
                "responses": {
                    "200": {
                        "description": "Dados do usuário autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "nome": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "nome": "a3fb6",
                                    "email": "teste@teste.com.br",
                                    "created_at": "2020-12-25T19:17:18.000000Z",
                                    "updated_at": "2020-12-25T19:17:18.000000Z"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Faltam informações"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Gateway",
            "description": "API Endpoints - Gateway"
        }
    ]
}